// pilotbank2-angular-master/src/app/models/registrationPg2.model.ts
export class RegistrationPg2 {
  selectedTitle: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  SIN: string;
  streetName: string;
  streetNumber: string;
  suitAppUnitNumber: string;
  province: string;
  cityTown: string;
  postalCode: string;
  industry: string;
  occupation: string;
}